refactor(service/logger): split sink/ into sink/ (transports) + middleware/ (combinators)
Every package in service/logger/sink/ implemented the same Sink
interface but played one of two very different roles:

- **Transports (leaf sinks)** adapt the logger to the outside world —
  they own an fd or a net.Conn and eventually call a syscall. The
  three packages here are console, file, syslog.

- **Middleware (Sink combinators)** take one or more Sinks and return
  an enriched Sink without touching the outside world themselves. The
  six packages here are async, failover, multi, recover, route, sample.

Lumping both families under sink/ made the directory read as "a pile
of sinks" and forced contributors to crack open each README to figure
out what each package actually did. The split surfaces the role in the
import path itself:

  internal/service/logger/
  ├── sink/                (3 leaf packages)
  │   ├── console/
  │   ├── file/
  │   └── syslog/
  └── middleware/          (6 combinator packages)
      ├── async/
      ├── failover/
      ├── multi/
      ├── recover/
      ├── route/
      └── sample/

Mechanics:
- git mv of 6 package directories preserves every file's history.
- Import paths rewritten across the affected .go files + string
  literals inside Private/Public error messages so the errs audit
  trail still points at the right code location.
- BUILD.bazel files regenerated via gazelle; pkg/v1/logger now
  imports multi from middleware/.
- docs/adr/0002 registry table updated so the code-range audit's
  source of truth tracks the new paths.
- pkg/v1/logger/README now documents the transport / middleware
  distinction under the "richer sinks" section.

No behavioural change. No public API rename. Every .claude/contexts
reference to the old paths is left untouched because those files are
historical planning documents that should reflect the state of the
world at the time they were written.
