+++ internal/service/logger/middleware/async/async_sink_external_test.go
@@ -10,7 +10,7 @@ import (
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
 	"github.com/kitsunium/sdk/internal/core/logger/level"
 	"github.com/kitsunium/sdk/internal/kernel/errs"
-	"github.com/kitsunium/sdk/internal/service/logger/sink/async"
+	"github.com/kitsunium/sdk/internal/service/logger/middleware/async"
 )
 
 // recordingSink captures Write/Flush/Close invocations for assertions.
