+++ internal/service/logger/middleware/failover/codes.go
