+++ internal/service/logger/middleware/failover/failover_sink_internal_test.go
