+++ internal/service/logger/middleware/multi/multi_external_test.go
@@ -8,7 +8,7 @@ import (
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
 	"github.com/kitsunium/sdk/internal/core/logger/level"
 	"github.com/kitsunium/sdk/internal/kernel/errs"
-	"github.com/kitsunium/sdk/internal/service/logger/sink/multi"
+	"github.com/kitsunium/sdk/internal/service/logger/middleware/multi"
 )
 
 // recordingSink captures every Write/Flush/Close call for assertions.
