+++ internal/service/logger/middleware/recover/BUILD.bazel
@@ -3,28 +3,27 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 package(default_visibility = ["//internal/service:service_consumers"])
 
 go_library(
-    name = "route",
+    name = "recover",
     srcs = [
         "codes.go",
         "errors.go",
-        "route_entry.go",
-        "router_sink.go",
+        "panic_value.go",
+        "recover_sink.go",
     ],
-    importpath = "github.com/kitsunium/sdk/internal/service/logger/sink/route",
+    importpath = "github.com/kitsunium/sdk/internal/service/logger/middleware/recover",
     visibility = [
         "//:__subpackages__",
         "//internal/service:service_consumers",
     ],
     deps = [
         "//internal/core/logger",
-        "//internal/core/logger/level",
         "//internal/kernel/errs",
     ],
 )
 
 alias(
     name = "go_default_library",
-    actual = ":route",
+    actual = ":recover",
     visibility = [
         "//:__subpackages__",
         "//internal/service:service_consumers",
@@ -32,13 +31,13 @@ alias(
 )
 
 go_test(
-    name = "route_test",
+    name = "recover_test",
     srcs = [
-        "route_entry_external_test.go",
-        "router_sink_external_test.go",
-        "router_sink_internal_test.go",
+        "panic_value_internal_test.go",
+        "recover_sink_external_test.go",
+        "recover_sink_internal_test.go",
     ],
-    embed = [":route"],
+    embed = [":recover"],
     deps = [
         "//internal/core/logger",
         "//internal/core/logger/level",
