+++ internal/service/logger/middleware/route/router_sink_external_test.go
@@ -8,7 +8,7 @@ import (
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
 	"github.com/kitsunium/sdk/internal/core/logger/level"
 	"github.com/kitsunium/sdk/internal/kernel/errs"
-	"github.com/kitsunium/sdk/internal/service/logger/sink/route"
+	"github.com/kitsunium/sdk/internal/service/logger/middleware/route"
 )
 
 type recordingSink struct {
