+++ internal/service/logger/middleware/sample/sample_sink_external_test.go
@@ -8,7 +8,7 @@ import (
 	corelogger "github.com/kitsunium/sdk/internal/core/logger"
 	"github.com/kitsunium/sdk/internal/core/logger/level"
 	"github.com/kitsunium/sdk/internal/kernel/errs"
-	"github.com/kitsunium/sdk/internal/service/logger/sink/sample"
+	"github.com/kitsunium/sdk/internal/service/logger/middleware/sample"
 )
 
 type countingSink struct {
